pdf-ai-poc/requirements.txt
openai>=1.40.0
pypdf>=4.3.1
python-dotenv>=1.0.1
